Set webpack global object to 'this' instead of 'window'
When webpack transpiles UMD code, it ends up looking something like

```
!function(t,e) {
    // module type checking
}(window, function() {
    // actual module colde
})
```

which means that even if the `window` variable is checked in the source code, it's still automatically used by webpack. This would prevent the transpiled code to be used anywhere that `window` isn't defined, e.g. in node environments. Thus, change the config to accept whatever `this` is in the global scope, i.e. window || self || global.