Move try-catch inside hook callback so the plugin handles the errors
Throwing an error inside the hook.tap() callback means that webpack catches the error and fails the build. Moving the try-catch inside the callback means that the MockRequests plugin catches and handles it instead. TODO: It might be a good idea to add an option for the user to cancel the build if this plugin fails.