Support functional 'entry' webpack config values
If the user passes in a function for the entry value, regardless of if it's a normal function or a promise (e.g. `entry: () => 'index.js'` or `entry: Promise.resolve('index.js')`), it is converted to a Promise within the `entryOption` hook. As such, add a check for if the hook's `entry` param is a function and process it as normal after the promise resolves.