First attempt at generating types.d.ts files for src/ and bin/ automatically
tsd-jsdoc is a JSDoc template system that takes the output from `jsdoc` and generates types.d.ts files from them. Though, bin/ and src/ will likely need to have their own types files to maintain a separation of concerns, not to mention it might confuse users to have both the API exports as well as the utility file exports in the same index.js file.