Add support for TypeScript in JSDoc
This will allow the '@typedef {import}' syntax for TS files. The better-docs package seems to be one of the better JSDoc plugins that supports TS, but I still think the docdash template/theme is nice, so don't change that part. Also, update the source.include(Pattern) to both include .ts files and the index.js for the '@module mock-requests' entry