Hide typedefs from nav pane
The typedefs only appear under the 'Global' heading, not under any of the namespaces/classes/etc. so they aren't all that useful. After adding the new 'npm' nav link, the nav pane now has a scrollbar when the typedefs are present. Since that makes the page a bit less pleasant to navigate through, hide the typedefs so that no scrollbar is shown and because displaying them is less helpful than it would be if they were also shown under their parents (as opposed to only being shown under 'Global')