Refactor code to allow 'slash URLs' to work with any origin
"Slash URLs" are URLs that begin with '/' instead of the domain/origin (e.g. `/myApi` instead of `https://website.com/myApi`). When using `fetch`/`XHR` directly, this poses no issue to MockRequests. But when using some third-party libraries, they automatically prepend a user-specified domain to the URL which breaks the `urlIsMocked()` check.

For example, if a user wrote a wrapper function around Axios where the parameter was just the API path without the domain, you'd end up with:

```javascript
function doGet(url = '/myApi') {
    axios.get(
        url,
        {
            baseUrl: 'https://website.com'
        }
    );
}

// will call `fetch('https://website.com/myApi')` instead of `fetch('/myApi')`
```

In order to fix this, the URL passed to `fetch`/`XHR` needs to be split up more than just the pathname and query parameters. If the user mocked `/myApi`, we need to be able to check if the pathname is mocked regardless of the domain/origin. This is especially the case when users have testing environments (which is the primary motivation for writing wrapper functions around e.g. Axios).

Thus, we do these primary changes:

* Split up the URL into more segments in `getPathnameAndQueryParams()` for more in-depth parsing.
* Normalize the URLs to remove ending slashes (to avoid e.g. `mock('/myApi')` not working for `fetch('/myApi/')`).
    - Involves changing what is stored in the `urlResponseMap`.
* Iterate through all possible URL combinations to find which combo was actually stored in the `urlResponseMap`, and then check if the URL from `fetch` matches them accordingly.
    - Do so in new `getMockedUrlFromUrlArg()` function.
    - This also does an additional `location.origin` check to see if the mocked URL was created from the actual origin vs another origin.
        + **TODO** This is likely unnecessary; consider removing it.
        + **TODO** Add tests for this if it is kept.
    - **TODO** More checks need to be added, see the TODO code blocks.

More TODOs:

* The code is very messy and has some browser-only tests in the src file.
* Clean up some repeated code.
* Plenty of others, see the code.