Add support for the 'XHR.onloadend' handler/event
It's another handler/event commonly used to check the status of an XHR request since it helps show progress of the request instead of just a boolean of whether or not it's finished.

It's also what Axios now prefers over onreadystatechange. See [this GitHub issue](https://github.com/D-Pow/MockRequests/issues/16) for more details.