Update app and demo to use new NodeJS versions
It is recommended to use NodeJS >= v14, but this app was last built with NodeJS v10. Thus, update both the app and demo with NodeJS v16. In the process, add typedefs for `jest` and `node` for better IDE autocompletion and replace the deprecated `node-sass` with `sass`. Note: The app and demo both still run perfectly on IE 11.