Add `publish` npm script for easier package publishing
NOTE: REQUIRES `npm run publish` INSTEAD OF `npm publish`!!

While package.json has a `files` field that allows specifying a particular directory in the published package, it doesn't offer a way to publish *only* the files in that dir, i.e. it will *keep* `dist/` in the published package (along with files always included: package.json, README, LICENSE, and the file in the `main` field) rather than only the files *within* `dist/` (see: https://docs.npmjs.com/cli/v7/configuring-npm/package-json#files).

After a short search, it seems the options we have are:

1. `cd dist && npm publish` - Simplest option.
2. `npm publish ./dist` - Doesn't work if inside a `publish` npm script (it would infinitely recurse b/c `./package.json` calls `publish` in `dist/package.json` which would try to call it in `dist/dist/package.json`), but does work if using a different name (e.g. `pub`, `release`, or similar).
3. Rewrite the library to make `main` point to `dist/index.js` as well as moving the Webpack plugin to a reasonable directory name (requires reconfiguring webpack.config.js, changing the dir name for the Webpack plugin, etc.).

For now, let's pick the first option because it's the simplest of the three. Eventually, we should implement the third so that we can just run `npm publish` in the root dir without needing a custom npm script.

Also, add a `prepublish` script to ensure the deployed package contains the latest code changes.

Refs:

* Using `cd dist && npm publish`:
    - https://stackoverflow.com/questions/48124806/npm-publish-a-package-with-only-the-children-of-dist-folder/48460033#48460033
    - https://stackoverflow.com/questions/37862712/how-to-publish-contents-only-of-a-specific-folder
* Using `npm publish ./dist`:
    - https://stackoverflow.com/questions/38935176/how-to-npm-publish-specific-folder-but-as-package-root/69646778#69646778