Ensure custom, mocked properties are always defined
Not sure if it was a new browser update or the Babel upgrade, but it seems that `Object.defineProperties(xhr, props)` doesn't seem to work for read-only properties, like `response`. Thus, ensure the mock values are always set by replacing `writable`/`value` with `get()`/`set()` since those properties tend to be getters with unset setters.